NIFI-12632 Extract SFTP components out of the standard bundle
