+++ nifi-nar-bundles/nifi-sftp-bundle/nifi-sftp-processors/src/main/java/org/apache/nifi/sftp/socket/ClientConfigurationException.java
@@ -1,0 +1,33 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.nifi.sftp.socket;
+
+import org.apache.nifi.processor.exception.ProcessException;
+
+/**
+ * Client Configuration Exception for specific failures during Client configuration
+ */
+public class ClientConfigurationException extends ProcessException {
+
+    public ClientConfigurationException(final String message) {
+        super(message);
+    }
+
+    public ClientConfigurationException(final String message, final Throwable cause) {
+        super(message, cause);
+    }
+}
