+++ nifi-nar-bundles/nifi-sftp-bundle/nifi-sftp-processors/src/main/resources/docs/org.apache.nifi.processors.standard.ListSFTP/ListSFTP-batch-processing.png
