+++ nifi-nar-bundles/nifi-standard-bundle/pom.xml
@@ -37,7 +37,6 @@
     <properties>
         <yammer.metrics.version>2.2.0</yammer.metrics.version>
         <jolt.version>0.1.8</jolt.version>
-        <org.apache.sshd.version>2.12.0</org.apache.sshd.version>
         <tika.version>2.9.1</tika.version>
     </properties>
     <dependencyManagement>
@@ -97,11 +96,6 @@
                 <groupId>com.yammer.metrics</groupId>
                 <artifactId>metrics-ganglia</artifactId>
                 <version>${yammer.metrics.version}</version>
-            </dependency>
-            <dependency>
-                <groupId>com.hierynomus</groupId>
-                <artifactId>sshj</artifactId>
-                <version>0.38.0</version>
             </dependency>
             <dependency>
                 <groupId>jakarta.activation</groupId>
